+++ examples/request-reply.php
@@ -21,7 +21,7 @@ use IDCT\NATS\Core\NatsMessage;
 
 use function Amp\async;
 
-$url = getenv('NATS_URL') ?: 'nats://127.0.0.1:14222';
+$url = getenv('NATS_URL') ?: 'nats://127.0.0.1:4222';
 
 $client = new NatsClient(new NatsOptions(servers: [$url], name: 'example-request-reply'));
 $client->connect()->await();